+++ Changelog.md
@@ -5,3 +5,6 @@
 - Adapted default position of some buttons
   - Please delete or manually update the config if something is broken
 - Updated icons
+
+## Fixes
+- Sorting wrong slots with Nemo's Backpacks loaded